+++ packages/asana/src/Adaptor.js
@@ -1,4 +1,3 @@
-/** @module Adaptor */
 import {
   execute as commonExecute,
   composeNextState,