+++ packages/dhis2/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/language-dhis2
 
+## 3.2.4
+
+### Patch Changes
+
+- 6d8de03: change @constructor to @function and remove /\*_ @module Adaptor _/
+
 ## 3.2.3
 
 ### Patch Changes