+++ packages/godata/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/language-godata
 
+## 3.2.1
+
+### Patch Changes
+
+- 6d8de03: change @constructor to @function and remove /\*_ @module Adaptor _/
+
 ## 3.2.0
 
 ### Minor Changes