+++ packages/mssql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/language-mssql
 
+## 4.0.1
+
+### Patch Changes
+
+- 6d8de03: change @constructor to @function and remove /\*_ @module Adaptor _/
+
 ## 4.0.0
 
 ### Major Changes