+++ packages/mysql/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/language-mysql
 
+## 1.3.1
+
+### Patch Changes
+
+- 6d8de03: change @constructor to @function and remove /\*_ @module Adaptor _/
+
 ## 1.3.0
 
 ### Minor Changes