+++ packages/openmrs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/language-openmrs
 
+## 0.10.1
+
+### Patch Changes
+
+- 6d8de03: change @constructor to @function and remove /\*_ @module Adaptor _/
+
 ## 0.10.0
 
 ### Minor Changes