+++ packages/sftp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @openfn/language-sftp
 
+## 0.6.5
+
+### Patch Changes
+
+- 6d8de03: change @constructor to @function and remove /\*_ @module Adaptor _/
+
 ## 0.6.4
 
 ### Patch Changes